+++ CHANGELOG.md
@@ -1,6 +1,9 @@
 I will record changes to this file just so i don't need to look at commit history every time
 
 # Changelog
+2025-10-04:
+- Created a script for updating linux on schedule
+
 2025-09-27:
 - Just created this file. Everything before this date is in commit history.
 - Added kinto.py config file to repo.